+++ lib/onlyoffice_mysql_helper.rb
@@ -2,6 +2,7 @@
 
 require 'onlyoffice_logger_helper'
 require 'onlyoffice_mysql_helper/mysql_helper'
+require 'onlyoffice_mysql_helper/mysql_logger'
 require 'onlyoffice_mysql_helper/version'
 
 # Helper class to work with MySQL