+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New Features
+
+* Add `MySQLLogger` class
+
 ### Changes
 
 * Rename `get_tables` to `tables`